requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.138
botocore==1.34.138
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.1
daphne==4.1.2
diff-match-patch==20230430
Django==5.1.3
django-cleanup==8.0.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-import-export==4.0.7
django-quill-editor==0.1.40
django-storages==1.14.3
django-timezone-field==6.1.0
django-webpack-loader==3.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==22.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
kombu==5.3.7
msgpack==1.1.0
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.45
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-crontab==3.1.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
sentry-sdk==2.6.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3